Telephony: Fix imis is not correct on android L for CT 3G sim card.
XSL-330

xPick below qcom changes:
------------------------------------------------------------------
Refactor IMSI and SubscriberId handling for CDMAPhone/CDMALTEPhone
IccRecords:
CdmaServiceStateTracker does not need to set IMSI on an IccRecord.
CDMAPhone:
When queried for the subscriber Id return the IMSI from
CdmaServiceStateTracker or the card depending on whether the active
subscription source is NV or RUIM respectively.

CRs-Fixed: 323977,323977,342146,505580
------------------------------------------------------------------

Change-Id: I062cd48ccf9b38b0ac6b56117f831e7155caa069
Signed-off-by: wangminfei <wangminfei@xiaomi.com>